+++ design/stories/package-list--item.stories.js
@@ -1,6 +1,3 @@
  
 export default {
-  title: "Components/Package List"
-};
-
-export const PackageListItem = () => "<li class=\"package-list-item\"><a href=\"/packages/@haskell/base\" class><h4 class=\"package-list-item__name\"><strong class>@haskell/base</strong></h4><p class=\"package-list-item__synopsis\">Basic libraries</p><div class=\"package-list-item__version\">4.16.0.0</div></a></li>"
+  title: "Com