+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Split the project into internal libraries ([#337](https://github.com/flora-pm/flora-server/pull/337)) 
 * Fetch and store deprecation information about packages ([#342](https://github.com/flora-pm/flora-server/pull/342))
 * Only index versionless package pages ([#343](https://github.com/flora-pm/flora-server/pull/343))
+* Display deprecation information on the package page ([#344](https://github.com/flora-pm/flora-server/pull/344))
 
 ## 1.0.9 -- 2023-01-06
 * Fix package title size in smaller screens ([#297](https://github.com/flora-pm/flora-server/pull/297))