+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Only index versionless package pages ([#343](https://github.com/flora-pm/flora-server/pull/343))
 * Display deprecation information on the package page ([#344](https://github.com/flora-pm/flora-server/pull/344))
 * Display deprecation information for releases ([#347](https://github.com/flora-pm/flora-server/pull/347))
+* Make package listings denser ([#355](https://github.com/flora-pm/flora-server/pull/355))
 
 ## 1.0.9 -- 2023-01-06
 * Fix package title size in smaller screens ([#297](https://github.com/flora-pm/flora-server/pull/297))